requirements.txt
apollo_torch==1.0.3
av==14.2.0
bitsandbytes==0.46.1
datasets==3.2.0
fastapi==0.116.1
fire==0.7.0
galore_torch==1.0
gradio==5.40.0
huggingface_hub==0.29.2
jieba==0.42.1
librosa==0.10.2.post1
matplotlib==3.10.5
MoD==0.3.0
modelscope==1.28.1
nltk==3.9.1
numpy==2.3.2
openai==1.98.0
openmind==1.0.0
packaging==25.0
pandas==2.3.1
peft==0.12.0
pillow==11.3.0
psutil==7.0.0
pydantic==2.11.7
pytest==8.3.5
PyYAML==6.0.2
ray==2.40.0
requests==2.32.4
rouge_chinese==1.0.3
safetensors==0.5.3
setuptools==75.1.0
sse_starlette==2.2.1
swanlab==0.6.8
torch==2.5.1
tqdm==4.66.5
transformers==4.49.0
trl==0.16.1
typing_extensions==4.14.1
uvicorn==0.35.0
vllm==0.7.3
wandb==0.21.0
svgwrite
cairosvg
IPython
six>=1.12.0
svglib
jupyter
scikit-image
networkx
moviepy
numba
sklearn
umap-learn
umap-learn[plot]
shapely
kivy